Allow users to upload files of products
Why:
We would like for users to be able to upload a csv of product data from
the products index.

This commit:
Introduces an upload csv button to the products page, a new imports
controller, and a class method on Product to handle the import logic. A
sample fixture was added for a feature test.