Refactor importer and raise if unknown file type
Why:
We would like to introduce dedicated service objects for importing
various file types. We would also like to raise if the importer is
provided a file type that is not know to the program

This commit:
factors out a new csv_importer and xlsx_importer from the
product_data_importer.